build(server): Update ts compiler && moving app files to src folder
Move app files to ./src

References:
- https://github.com/Microsoft/TypeScript-Node-Starter#configuring-typescript-compilation
- https://www.typescriptlang.org/tsconfig
