+++ server/package.json
@@ -6,11 +6,14 @@
   "engines": {
     "node": "14.x"
   },
-  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon app.ts",
-    "postinstall": "cd .. && husky install client/.husky"
+    "postinstall": "cd .. && husky install client/.husky",
+    "start": "tsc && node dist/index.js",
+    "dev": "concurrently -k -n \"Typescript,Node\" -p \"[{name}]\" -c \"cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run watch-node\"",
+    "postinstall": "cd .. && npx husky install client/.husky",
+    "watch-ts": "tsc -w",
+    "watch-node": "nodemon dist/index.js"
   },
   "eslintConfig": {
     "extends": [
@@ -27,6 +30,7 @@
   "devDependencies": {
     "@types/express": "^4.17.9",
     "@types/node": "^14.14.14",
+    "concurrently": "^5.3.0",
     "eslint": "^7.15.0",
     "eslint-config-prettier": "^7.0.0",
     "eslint-plugin-prettier": "^3.3.0",
