+++ server/tsconfig.json
@@ -1,13 +1,18 @@
 {
   "compilerOptions": {
-    "module": "commonjs",
+    "module": "CommonJS",
     "moduleResolution": "node",
+    "esModuleInterop": true,
+    // "strict": true,
     "pretty": true,
     "sourceMap": true,
-    "target": "es6",
-    "outDir": "./dist",
-    "baseUrl": "./lib"
+    "target": "ES2020",
+    "outDir": "dist",
+    "baseUrl": ".",
+    "paths": {
+      "*": ["node_modules/*", "src/types/*"]
+    }
   },
-  "include": ["**/*.ts"],
+  "include": ["src/**/*"],
   "exclude": ["node_modules"]
 }
